move create-test-store
