+++ packages/@coorpacademy-app-review/src/actions/test/create-test-store.ts
@@ -2,8 +2,8 @@ import type {ExecutionContext} from 'ava';
 import constant from 'lodash/fp/constant';
 import {AnyAction, applyMiddleware, compose, createStore, Dispatch, Middleware, Store} from 'redux';
 import thunk from 'redux-thunk';
-import rootReducer, {StoreState} from '../../../reducers';
-import {Services} from '../../../types/common';
+import rootReducer, {StoreState} from '../../reducers';
+import {Services} from '../../types/common';
 
 const assertActionsMiddleware = (t: ExecutionContext, ACTIONS: AnyAction[]): Middleware =>
   constant((next: Dispatch) => (action: AnyAction): unknown => {
